pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "bluefusion"
version = "0.1.0"
description = "AI-powered dual BLE interface controller for MacBook and BLE sniffers"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "BlueFusion Team", email = "bluefusion@example.com"},
]
dependencies = [
    "bleak>=0.21.0",
    "pyserial>=3.5",
    "asyncio>=3.4.3",
    "pydantic>=2.0",
    "rich>=13.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "aiofiles>=23.0",
    "structlog>=23.0",
    "click>=8.1.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "websockets>=11.0",
    "gradio>=4.0.0",
    "httpx>=0.25.0",
    "plotly>=5.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "httpx>=0.25.0",
    "pytest-mock>=3.12.0",
]
ai = [
    "scikit-learn>=1.3.0",
    "tensorflow>=2.13.0",
    "transformers>=4.30.0",
]

[project.scripts]
bluefusion = "bluefusion:cli"

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W", "B", "Q"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--tb=short"
testpaths = ["tests"]